UI Design: initial design to print music sheet lines
in this commit, I made the drawMeasure.fxml to create one measure of
music sheet. I added the update method to the previewMusic.java to be
used in the previewButtonHandler method in the mainViewController class
to update the music sheet. I also modified the previewMusic.fxml to
allow viewing the measures.