UI: update the draw class/ preview.Java
updated the MLine class to create lines and return specific starting x
and y value based on given tag.
updated DrawMusicLine class to drawing music lines at specific position.
each music line (group of 6 lines) are also added to a list so we can
use the `string` attributed extracted by the parser to locate the
position of notes.
finally added the code that allows to draw the music lines and notes on
the preview window.