Made classes for drawing music and bar lines for drums.
Since drums will use 5 music lines (instead of 6 music lines for guitar,
where each line is a guitar string), the drumset sheet music now gets
drawn with 5 lines. There is still a 6th line at the top (which is there
to simplify the drawing later), but it is invisible (white colour).