Created test class for DrawNote and added test for drawFret method.
Created JUnit test class for the DrawNote class and added a test method
for testing the drawFret method. This converts a tablature to a score,
then draws all the notes in the score to a pane, checking the pane after
every note to make sure it was positioned correctly and has the correct
text value.