Added test method in DrawNoteTest for drawO and drawX methods.
Added test methods in the DrawNoteTest JUnit test class to test the
drawO and drawX methods of the DrawNote class. This converts a tablature
to a score and draws all the notes to a pane. After each note is added,
the pane is checked to make sure the note was given the correct x- and
y-positions and the correct text value ("o" for drawO or "x" for drawX).