Updated JUint test case for drawing drum notes.
Since beams are being drawn, the previous way of testing drum note
drawing does not work. Updated this test case to count the number of
notes drawn instead. This is done by counting the notes that have a
particular id corresponding to their note type (this meant changing the
DrawDrumsetNote class to give each note an id).